feat: restructure SEO components and metadata handling
- Introduced a new SEO module with structured exports for better organization.
- Added JSON-LD generation functions for various schema types including articles, organizations, and websites.
- Implemented metadata building functions to enhance page SEO with Open Graph and Twitter card support.
- Refactored blog and index pages to utilize new components for improved readability and maintainability.
- Removed deprecated image loader tests and unused type definitions.
