+++ src/components/pages/PublicationsClient.tsx
@@ -16,6 +16,7 @@ import {
   toggleSetValue,
 } from '@/components/filters/filterHelpers';
 import {
+  formatClearFilterLabel,
   formatFilterResultCount,
   formatRemoveFilterAriaLabel,
   resolveFilterText,
@@ -41,6 +42,41 @@ type Item = {
 };
 
 const publicationTypeOrder: Item['type'][] = ['paper', 'app', 'article', 'talk', 'slide', 'media'];
+
+const publicationTypeLabels: Record<Locale, Record<Item['type'], string>> = {
+  ja: {
+    paper: '📄 論文',
+    article: '📝 技術ブログ',
+    talk: '🎤 登壇',
+    slide: '📑 スライド',
+    media: '📰 メディア',
+    app: '📱 アプリ',
+  },
+  en: {
+    paper: '📄 Papers',
+    article: '📝 Technical Articles',
+    talk: '🎤 Talks',
+    slide: '📑 Slides',
+    media: '📰 Media',
+    app: '📱 Apps',
+  },
+  zh: {
+    paper: '📄 论文',
+    article: '📝 技术博客',
+    talk: '🎤 演讲',
+    slide: '📑 幻灯片',
+    media: '📰 媒体',
+    app: '📱 应用',
+  },
+  fr: {
+    paper: '📄 Articles scientifiques',
+    article: '📝 Articles techniques',
+    talk: '🎤 Presentations',
+    slide: '📑 Slides',
+    media: '📰 Medias',
+    app: '📱 Apps',
+  },
+};
 
 function isPublicationType(value: string): value is Item['type'] {
   return publicationTypeOrder.includes(value as Item['type']);
@@ -162,40 +198,7 @@ export function PublicationsClient({ items, locale = 'en' }: { items: Item[]; lo
   }, [availableTypes, groups, q, sort, typeFiltered]);
 
   const t = resolveFilterText(locale);
-  const typeLabels: Record<Item['type'], string> = {
-    ja: {
-        paper: '📄 論文',
-        article: '📝 技術ブログ',
-        talk: '🎤 登壇',
-        slide: '📑 スライド',
-        media: '📰 メディア',
-        app: '📱 アプリ',
-      },
-    en: {
-        paper: '📄 Papers',
-        article: '📝 Technical Articles',
-        talk: '🎤 Talks',
-        slide: '📑 Slides',
-        media: '📰 Media',
-        app: '📱 Apps',
-      },
-    zh: {
-      paper: '📄 论文',
-      article: '📝 技术博客',
-      talk: '🎤 演讲',
-      slide: '📑 幻灯片',
-      media: '📰 媒体',
-      app: '📱 应用',
-    },
-    fr: {
-      paper: '📄 Articles scientifiques',
-      article: '📝 Articles techniques',
-      talk: '🎤 Presentations',
-      slide: '📑 Slides',
-      media: '📰 Medias',
-      app: '📱 Apps',
-    },
-  }[locale];
+  const typeLabels = publicationTypeLabels[locale];
   const resultLabel = useMemo(
     () => formatFilterResultCount(locale, typeFiltered.length, items.length),
     [items.length, locale, typeFiltered.length],
@@ -240,7 +243,7 @@ export function PublicationsClient({ items, locale = 'en' }: { items: Item[]; lo
     if (selectedTypeSet.size !== availableTypes.length) {
       actions.push({
         key: 'clear-types',
-        label: locale === 'ja' ? `${t.types}をクリア` : formatClearTypesLabel(locale, t.types),
+        label: formatClearFilterLabel(locale, t.types),
         onClick: () => setSelectedTypes(null),
       });
     }
@@ -420,10 +423,4 @@ export function PublicationsClient({ items, locale = 'en' }: { items: Item[]; lo
       )}
     </div>
   );
-}
-
-function formatClearTypesLabel(locale: Locale, label: string) {
-  if (locale === 'zh') return `清除${label}`;
-  if (locale === 'fr') return `Effacer ${label}`;
-  return `Clear ${label}`;
 }
