--- src/types/squoosh-lib.d.ts
@@ -1,8 +1,0 @@
-declare module '@squoosh/lib' {
-  export class ImagePool {
-    constructor(concurrency?: number);
-    ingestImage(input: ArrayBuffer | Uint8Array | Buffer): any;
-    close(): Promise<void>;
-  }
-}
-
